feat: rename plugin data to export-data
