+++ src/services/chat/repository.ts
@@ -5,8 +5,8 @@ import { ResultOf } from '@graasp/sdk';
 import { AppDataSource } from '../../plugins/datasource';
 import { MemberIdentifierNotFound } from '../itemLogin/errors';
 import { Member } from '../member/entities/member';
-import { messageSchema } from '../member/plugins/data/schemas/schemas';
-import { schemaToSelectMapper } from '../member/plugins/data/utils/selection.utils';
+import { messageSchema } from '../member/plugins/export-data/schemas/schemas';
+import { schemaToSelectMapper } from '../member/plugins/export-data/utils/selection.utils';
 import { mapById } from '../utils';
 import { ChatMessage } from './chatMessage';
 import { ChatMessageNotFound } from './errors';
