+++ src/services/item/plugins/app/appAction/repository.ts
@@ -4,8 +4,8 @@ import { ResultOf } from '@graasp/sdk';
 
 import { AppDataSource } from '../../../../../plugins/datasource';
 import { MemberIdentifierNotFound } from '../../../../itemLogin/errors';
-import { appActionSchema } from '../../../../member/plugins/data/schemas/schemas';
-import { schemaToSelectMapper } from '../../../../member/plugins/data/utils/selection.utils';
+import { appActionSchema } from '../../../../member/plugins/export-data/schemas/schemas';
+import { schemaToSelectMapper } from '../../../../member/plugins/export-data/utils/selection.utils';
 import { mapById } from '../../../../utils';
 import { ManyItemsGetFilter, SingleItemGetFilter } from '../interfaces/request';
 import { AppAction } from './appAction';
