+++ src/services/item/plugins/app/appSetting/repository.ts
@@ -3,8 +3,8 @@ import { ItemType } from '@graasp/sdk';
 import { AppDataSource } from '../../../../../plugins/datasource';
 import { ItemNotFound } from '../../../../../utils/errors';
 import { MemberIdentifierNotFound } from '../../../../itemLogin/errors';
-import { appSettingSchema } from '../../../../member/plugins/data/schemas/schemas';
-import { schemaToSelectMapper } from '../../../../member/plugins/data/utils/selection.utils';
+import { appSettingSchema } from '../../../../member/plugins/export-data/schemas/schemas';
+import { schemaToSelectMapper } from '../../../../member/plugins/export-data/utils/selection.utils';
 import { AppSetting } from './appSettings';
 import { AppSettingNotFound, PreventUpdateAppSettingFile } from './errors';
 import { InputAppSetting } from './interfaces/app-setting';
