+++ src/services/item/plugins/itemCategory/repositories/itemCategory.ts
@@ -3,8 +3,8 @@ import { QueryFailedError } from 'typeorm';
 import { AppDataSource } from '../../../../../plugins/datasource';
 import { MemberNotFound } from '../../../../../utils/errors';
 import { DUPLICATE_ENTRY_ERROR_CODE } from '../../../../../utils/typeormError';
-import { itemCategorySchema } from '../../../../member/plugins/data/schemas/schemas';
-import { schemaToSelectMapper } from '../../../../member/plugins/data/utils/selection.utils';
+import { itemCategorySchema } from '../../../../member/plugins/export-data/schemas/schemas';
+import { schemaToSelectMapper } from '../../../../member/plugins/export-data/utils/selection.utils';
 import { Item } from '../../../entities/Item';
 import { ItemCategory } from '../entities/ItemCategory';
 import { DuplicateItemCategoryError } from '../errors';
