+++ src/services/item/plugins/itemFavorite/repositories/favorite.ts
@@ -3,8 +3,8 @@ import { EntityManager, Repository } from 'typeorm';
 import { AppDataSource } from '../../../../../plugins/datasource';
 import { DUPLICATE_ENTRY_ERROR_CODE } from '../../../../../utils/typeormError';
 import { MemberIdentifierNotFound } from '../../../../itemLogin/errors';
-import { itemFavoriteSchema } from '../../../../member/plugins/data/schemas/schemas';
-import { schemaToSelectMapper } from '../../../../member/plugins/data/utils/selection.utils';
+import { itemFavoriteSchema } from '../../../../member/plugins/export-data/schemas/schemas';
+import { schemaToSelectMapper } from '../../../../member/plugins/export-data/utils/selection.utils';
 import { Item } from '../../../entities/Item';
 import { ItemFavorite } from '../entities/ItemFavorite';
 import { DuplicateFavoriteError, ItemFavoriteNotFound } from '../errors';
