+++ src/services/itemMembership/repository.ts
@@ -22,8 +22,8 @@ import { Item } from '../item/entities/Item';
 import { ItemSearchParams, Ordering, SortBy } from '../item/types';
 import { MemberIdentifierNotFound } from '../itemLogin/errors';
 import { Member } from '../member/entities/member';
-import { itemMembershipSchema } from '../member/plugins/data/schemas/schemas';
-import { schemaToSelectMapper } from '../member/plugins/data/utils/selection.utils';
+import { itemMembershipSchema } from '../member/plugins/export-data/schemas/schemas';
+import { schemaToSelectMapper } from '../member/plugins/export-data/utils/selection.utils';
 import { mapById } from '../utils';
 import { ItemMembership } from './entities/ItemMembership';
 import { getPermissionsAtItemSql } from './utils';
